requirements.txt
GPUtil
absl-py==2.1.0
aio-pika==9.4.3
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiormq==6.8.1
aiosignal==1.3.1
alembic==1.13.3
array-record==0.5.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
autopage==0.5.2
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cliff==4.7.0
cmaes==0.11.1
cmd2==2.4.3
colorlog==6.8.2
contourpy==1.3.0
cycler==0.12.1
dataclasses==0.6
dataclasses-json==0.6.7
dm-tree==0.1.8
etils==1.5.2
flask==3.0.3
flask-socketio==5.4.1
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
gevent
google-pasta==0.2.0
greenlet==1.1.3.post0
grpcio==1.66.2
h11==0.14.0
h5py==3.12.1
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.2
joblib==1.4.2
keras==3.5.0
keras-preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
mako==1.3.5
markdown==3.7
markdown-it-py==3.0.0
markupsafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
numpy==1.26.4
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu12==10.3.4.107
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu12==12.2.0.103
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
opencv-python==4.10.0.84
opt-einsum==3.4.0
optree==0.13.0
optuna==2.3.0
packaging==24.1
pamqp==3.3.0
pandas==2.2.3
pbr==6.1.0
pika==1.3.2
pillow==11.0.0
prettytable==3.11.0
promise==2.3
protobuf
psutil==6.0.0
pygments==2.18.0
pyparsing==3.2.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.2
pyyaml==6.0.2
requests==2.32.3
rich==13.9.1
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
simple-websocket==1.0.0
six==1.16.0
sqlalchemy==2.0.35
stevedore==5.3.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-datasets==4.9.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.0
termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
tqdm==4.66.5
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
yarl==1.13.1
zipp==3.20.2
zope-event==5.0
zope-interface==7.0.3
tabulate